Presentations of the different subteams that worked on the project
These are the different presentations of the subteams that worked on the projected and presented at the SASE Labs Symposium. The whole project was built under the leadership and supervision of me and my Co-Director and I specifically worked extensively on the Vision part of the project and the details can be seen more in the "Vision Team Symposium Presentation.pdf" presentation.